seperation route and controller for user
